+++ include/dxc/DXIL/DxilInstructions.h
@@ -11137,5 +11137,44 @@ struct DxilInst_MatrixOuterProduct {
   llvm::Value *get_vectorB() const { return Instr->getOperand(3); }
   void set_vectorB(llvm::Value *val) { Instr->setOperand(3, val); }
 };
+
+/// This instruction triggers a breakpoint if a debugger is attached
+struct DxilInst_DebugBreak {
+  llvm::Instruction *Instr;
+  // Construction and identification
+  DxilInst_DebugBreak(llvm::Instruction *pInstr) : Instr(pInstr) {}
+  operator bool() const {
+    return hlsl::OP::IsDxilOpFuncCallInst(Instr, hlsl::OP::OpCode::DebugBreak);
+  }
+  // Validation support
+  bool isAllowed() const { return true; }
+  bool isArgumentListValid() const {
+    if (1 != llvm::dyn_cast<llvm::CallInst>(Instr)->getNumArgOperands())
+      return false;
+    return true;
+  }
+  // Metadata
+  bool requiresUniformInputs() const { return false; }
+};
+
+/// This instruction returns true if a debugger is attached
+struct DxilInst_IsDebuggerPresent {
+  llvm::Instruction *Instr;
+  // Construction and identification
+  DxilInst_IsDebuggerPresent(llvm::Instruction *pInstr) : Instr(pInstr) {}
+  operator bool() const {
+    return hlsl::OP::IsDxilOpFuncCallInst(Instr,
+                                          hlsl::OP::OpCode::IsDebuggerPresent);
+  }
+  // Validation support
+  bool isAllowed() const { return true; }
+  bool isArgumentListValid() const {
+    if (1 != llvm::dyn_cast<llvm::CallInst>(Instr)->getNumArgOperands())
+      return false;
+    return true;
+  }
+  // Metadata
+  bool requiresUniformInputs() const { return false; }
+};
 // INSTR-HELPER:END
 } // namespace hlsl